+++ examples/nogallery/monitoring_plots.py
@@ -21,7 +21,7 @@ import seaborn as sns
 import pandas as pd
 import numpy as np
 
-from km3pipe import Pipeline, Module, Geometry
+from km3pipe import Pipeline, Module, Calibration
 from km3pipe.dataclasses import HitSeries
 from km3pipe.common import StringIO, Queue, Empty
 from km3pipe.hardware import Detector
@@ -42,8 +42,8 @@ from km3pipe.logger import logging
 PLOTS_PATH = '/home/km3net/monitoring/www/plots'
 N_DOMS = 18
 N_DUS = 2
-geometry = Geometry(det_id=14)
-detector = geometry.detector
+cal = Calibration(det_id=14)
+detector = cal.detector
 
 xfmt = md.DateFormatter('%Y-%m-%d %H:%M')
 lock = threading.Lock()
@@ -359,7 +359,7 @@ class ZTPlot(Module):
 
     def create_plot(self, event_info, hits):
         print(self.__class__.__name__ + ": updating plot.")
-        geometry.apply(hits)
+        cal.apply(hits)
         dus = set(detector.doms[h.dom_id][0] for h in hits)
 
         n_plots = len(dus)