+++ km3modules/topology.py
@@ -16,7 +16,7 @@ class TriggeredDUs(Module):
     """Check for triggered DUs."""
     def process(self, blob):
         triggered_hits = blob['Hits'].triggered_hits.serialise(to='pandas')
-        self.geometry.apply(triggered_hits)
+        self.calibration.apply(triggered_hits)
         dus = np.unique(triggered_hits['du'])
         n_dus = len(dus)